search,sort..

+++ src/main/webapp/WEB-INF/view/template/table_approve.jsp
@@ -6,9 +6,9 @@
         <thead>
             <tr>
                 <th data-th="Driver details"><span>STT</span></th>
-                <th data-th="Driver details"><a href="/admin?orderBy=id_user"> <span>${messageSource.getMessage("td.author",null,locale)}</span></a></th>
-                <th class="text-center"><a href="/admin?orderBy=title">${messageSource.getMessage("td.title",null,locale)}</a></th>
-                <th><a href="/admin?orderBy=time_post">${messageSource.getMessage("td.timePost",null,locale)}</a></th>
+                <th data-th="Driver details"><a href="/admin?orderBy=id_user"><img src="<s:url value="public/asserts/images/sort.png" />"> <span class="dp-inline">${messageSource.getMessage("td.author",null,locale)}</span></a></th>
+                <th class="text-center"><a href="/admin?orderBy=title"><img src="<s:url value="public/asserts/images/sort.png" />" class="mgr-5">${messageSource.getMessage("td.title",null,locale)}</a></th>
+                <th><a href="/admin?orderBy=time_post"><img src="<s:url value="public/asserts/images/sort.png" />" class="mgr-5">${messageSource.getMessage("td.timePost",null,locale)}</a></th>
                 <th>${messageSource.getMessage("td.action",null,locale)}</th>
             </tr>
         </thead>
@@ -23,7 +23,7 @@
                     <%--<a href="javascript:void(0)" onclick="A.getPostImprove('/admin-post-approve?action=approve&id='+${post.id},null)"> <span class="glyphicon glyphicon-ok mgr-10"></span></a>--%>
                     <%--<a href="javascript:void(0)" onclick="A.getPostImprove('/admin-post-approve?action=delete&id='+${post.id},null)"> <span class="glyphicon glyphicon-remove mgl-10"></span></a>--%>
                     <a href="/admin?page=${requestScope.page}&action=approve&id=${post.id}" > <span class="glyphicon glyphicon-ok mgr-10"></span></a>
-                    <a href="/admin?page=${requestScope.page}&action=delete&id=${post.id}"> <span class="glyphicon glyphicon-remove mgl-10"></span></a>
+                    <a href="/admin?page=${requestScope.page}&action=delete&id=${post.id}" onclick="return window.confirm('Are you sure you want to delete this post?')"> <span class="glyphicon glyphicon-remove mgl-10"></span></a>
                     <a href="/update?action=update&id=${post.id}"><img class="mgt--5 mgl-10" src="<s:url value="public/asserts/images/edit.gif"/>" alt=""></a>
 
                 </td>
@@ -35,7 +35,7 @@
     <div>
 
         <ul class="pager">
-            <jsp:useBean id="numberView" class="Utils.NumberViewSort"/>
+            <jsp:useBean id="numberView" class="utils.NumberViewSort"/>
             <c:if test="${requestScope.querySearch==null}">
                 <c:if test="${requestScope.page>=2}">
                     <li class="previous">
